+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # BSCAMPP v1.0.6
 1. Implemented functionality to recompile C++ codes locally so that dynamic
 libraries are referred to correctly. 
+2. Added UserWarning suppression to suppress warnings during subtree unifurcation
+suppression.
 
 # BSCAMPP v1.0.5
 1. Changed from using original sequence names from query input to using