+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# BSCAMPP v1.0.8 (WIP)
+1. Fixed the path issue with recompiling binaries on macOS.
+
 # BSCAMPP v1.0.7
 1. Fixed subtree formatting with pplacer output and the final output formatting
 regarding the difference between "n" and "nm" in jplace.